+++ changelog.md
@@ -4,6 +4,11 @@ Features
 
 -  Rebuild the commands with the newest unstable release
 
+Bugfix
+
+-  Fixed command command returning the wrong arity due to a Redis bug
+-  Fixed brpop command returning the wrong keystop due to a Redis bug
+
 ## v.1.1.0 - 09 Feb, 2016
 
 Features